+++ src/components/apply/ApplyForm/ApplyForm.component.tsx
@@ -1,4 +1,3 @@
-import { applicationApiService } from '@/api/services';
 import {
   LoadingModal,
   PersonalInformation,
@@ -9,17 +8,14 @@ import {
   BackToListLink,
   SubmittedButton,
   PersonalInfoAgree,
+  TempSaveAndSubmitButton,
 } from '@/components';
 import { PATH_NAME } from '@/constants';
 import { Application } from '@/types/dto';
-import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/router';
-import { MouseEventHandler, MutableRefObject, useRef, useState } from 'react';
+import { MutableRefObject, useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
-import {
-  ApplyFormValues,
-  APPLY_FORM_KEYS,
-} from '../PersonalInformation/PersonalInformation.component';
+import { ApplyFormValues } from '../PersonalInformation/PersonalInformation.component';
 import * as Styled from './ApplyForm.styled';
 
 interface ApplyFormProps {
@@ -28,7 +24,6 @@ interface ApplyFormProps {
 }
 
 const ApplyForm = ({ application, isSubmitted }: ApplyFormProps) => {
-  const session = useSession();
   const router = useRouter();
 
   const tempSaveButtonRef = useRef() as MutableRefObject<HTMLButtonElement>;
@@ -58,66 +53,8 @@ const ApplyForm = ({ application, isSubmitted }: ApplyFormProps) => {
   const applyForm = useForm<ApplyFormValues>();
   const {
     handleSubmit,
-    watch,
-    trigger,
-    setFocus,
     formState: { isDirty },
   } = applyForm;
-
-  const handleTempSaveApplication: MouseEventHandler<HTMLButtonElement> = async () => {
-    if (session.status === 'unauthenticated') return;
-
-    const { userName, phone, isAgreePersonalInfo, birthdate, department, residence } = watch();
-
-    if (!(await trigger(APPLY_FORM_KEYS.userName))) {
-      setFocus(APPLY_FORM_KEYS.userName);
-      return;
-    }
-
-    if (!(await trigger(APPLY_FORM_KEYS.phone))) {
-      setFocus(APPLY_FORM_KEYS.phone);
-      return;
-    }
-
-    if (!(await trigger(APPLY_FORM_KEYS.birthdate))) {
-      setFocus(APPLY_FORM_KEYS.birthdate);
-      return;
-    }
-
-    const updateApplicationRequest = {
-      applicantName: userName,
-      phoneNumber: phone,
-      privacyPolicyAgreed: isAgreePersonalInfo,
-      birthdate,
-      department,
-      residence,
-      answers: questionsAndAnswers.map(({ question, answer }) => {
-        const uniqueQuestionId = `question-${question.questionId}`;
-
-        return {
-          answerId: answer.answerId,
-          questionId: question.questionId,
-          content: watch(uniqueQuestionId),
-        };
-      }),
-    };
-
-    setIsRequesting(true);
-    try {
-      await applicationApiService.tempSaveApplication({
-        accessToken: session.data?.accessToken,
-        applicationId,
-        updateApplicationRequest,
-      });
-
-      setIsRequesting(false);
-      setIsOpenTempSaveSuccessAlertModal(true);
-      setIsTempSaved(true);
-    } catch (error) {
-      setIsRequesting(false);
-      setIsOpenTempSaveFailedAlertModal(true);
-    }
-  };
 
   const handleOpenSubmitModal = () => {
     setIsOpenConfirmSubmittedModal(true);
@@ -150,22 +87,16 @@ const ApplyForm = ({ application, isSubmitted }: ApplyFormProps) => {
               isCurrentlyOnDetailPage={isCurrentlyOnDetailPage}
             />
           ) : (
-            <>
-              <Styled.TempSaveButton
-                type="button"
-                disabled={!watch(APPLY_FORM_KEYS.isAgreePersonalInfo)}
-                onClick={handleTempSaveApplication}
-                ref={tempSaveButtonRef}
-              >
-                임시저장하기
-              </Styled.TempSaveButton>
-              <Styled.SubmitButton
-                disabled={!watch(APPLY_FORM_KEYS.isAgreePersonalInfo)}
-                ref={submitButtonRef}
-              >
-                제출하기
-              </Styled.SubmitButton>
-            </>
+            <TempSaveAndSubmitButton
+              applicationId={applicationId}
+              applyForm={applyForm}
+              questionsAndAnswers={questionsAndAnswers}
+              setIsOpenTempSaveFailedAlertModal={setIsOpenTempSaveFailedAlertModal}
+              setIsOpenTempSaveSuccessAlertModal={setIsOpenTempSaveSuccessAlertModal}
+              setIsRequesting={setIsRequesting}
+              setIsTempSaved={setIsTempSaved}
+              refs={{ submitButtonRef, tempSaveButtonRef }}
+            />
           )}
           <BackToListLink application={application} />
         </Styled.ControlSection>
