+++ src/components/apply/ApplyForm/ApplyForm.styled.ts
@@ -11,28 +11,3 @@ export const ControlSection = styled.div`
     }
   `}
 `;
-
-export const TempSaveButton = styled.button`
-  ${({ theme }) => css`
-    ${theme.button.type.primaryLine}
-    width: 20rem;
-    margin-right: 1.6rem;
-
-    @media (max-width: ${theme.breakPoint.media.mobile}) {
-      width: 100%;
-      margin-right: 0;
-      margin-bottom: 1.6rem;
-    }
-  `}
-`;
-
-export const SubmitButton = styled.button`
-  ${({ theme }) => css`
-    ${theme.button.type.primary}
-    width: 20rem;
-    @media (max-width: ${theme.breakPoint.media.mobile}) {
-      width: 100%;
-      margin-right: 0;
-    }
-  `}
-`;
